@@ -15,21 +15,24 @@
     "bootstrap": "^4.6.0",
     "css-loader": "^6.2.0",
     "date-fns": "^2.23.0",
+    "eslint": "^8.56.0",
+    "eslint-plugin-react": "^7.33.2",
     "file-loader": "^6.2.0",
     "isotope-layout": "^3.0.6",
-    "react": "^17.0.2",
+    "react": "^18.3.1",
     "react-awesome-reveal": "^3.8.1",
     "react-bootstrap": "^1.6.1",
     "react-bootstrap-icons": "^1.5.0",
-    "react-dom": "^17.0.2",
-    "react-redux": "^8.0.5",
-    "react-router-bootstrap": "^0.25.0",
-    "react-router-dom": "^5.2.0",
+    "react-dom": "^18.3.1",
+    "react-redux": "^8.1.3",
+    "react-router-bootstrap": "^0.26.2",
+    "react-router-dom": "^6.23.0",
     "sass": "^1.62.1",
     "sass-loader": "^12.1.0",
     "style-loader": "^3.2.1",
     "styled-components": "^6.0.0-rc.1",
     "typewriter-effect": "^2.19.0",
+    "url-loader": "^4.1.1",
     "webpack": "^5.46.0",
     "webpack-bundle-tracker": "^1.2.0",
     "webpack-cli": "^4.7.2"
@@ -39,5 +42,11 @@
     "Webpack"
   ],
   "author": "Eron Salling",
-  "license": "ISC"
+  "license": "ISC",
+  "dependencies": {
+    "@reduxjs/toolkit": "^1.9.7",
+    "axios": "^1.6.7",
+    "framer-motion": "^11.1.7",
+    "socket.io-client": "^4.7.2"
+  }
 }
