@@ -1,27 +1,25 @@
-import React from 'react'
-import { Image, Row, Col, Badge } from 'react-bootstrap'
-import { Github, Linkedin } from 'react-bootstrap-icons'
+import React from 'react';
+import PropTypes from 'prop-types';
+import { Image, Row, Col } from 'react-bootstrap';
+import { Github, Linkedin } from 'react-bootstrap-icons';
 
-import BadgeLink from '../shared/BadgeLink'
+import BadgeLink from '../shared/BadgeLink';
+import PageContainer from '../shared/PageContainer';
 
-const About = () => (
-	<section id="about">
+const About = ({
+	description,
+	img,
+}) => (
+	<PageContainer id="about">
 		<Row className="justify-content-center">
 			<Col className="pb-3" md={6} sm={12}>
 				<Image
-					src="https://eron-portfolio.s3.amazonaws.com/Eron+Headshot+Big"
+					src={img}
 					fluid
 				/>
 			</Col>
 			<Col md={6} sm={12}>
-				<p>
-					Hey there, I'm Eron Salling (he/him). I'm a game developer who
-					specializes in educational experiences and civic engagement.
-				</p>
-				<p>
-					When building games I tend to lean on ReactJS, along with other frontend tools like Redux, Socket.io, and more.
-					As a full-stack developer I've worked on many stacks and consider myself a polyglot, specializing in JavaScript, Python, and C# development. 
-				</p>
+				<p>{description}</p>
 
 				<div>
 					<BadgeLink variant="primary" url="https://github.com/esalling23">
@@ -33,7 +31,12 @@ const About = () => (
 				</div>
 			</Col>
 		</Row>
-	</section>
-)
+	</PageContainer>
+);
 
-export default About
+About.propTypes = {
+	description: PropTypes.string,
+	img: PropTypes.string
+};
+
+export default About;
