@@ -1,7 +1,24 @@
 from django.contrib import admin
+from import_export import resources
+from import_export.admin import ImportExportModelAdmin
 
 from .models import Project, Category
 
-# Register your models here.
-admin.site.register(Project)
-admin.site.register(Category)
+# Import/export resource classes
+class ProjectResource(resources.ModelResource):
+		class Meta: 
+				model = Project
+
+class CategoryResource(resources.ModelResource):
+		class Meta: 
+				model = Category
+
+# Custom Admin handling
+class ProjectAdmin(ImportExportModelAdmin):
+		resource_classes = [ProjectResource]
+
+class CategoryAdmin(ImportExportModelAdmin):
+		resource_classes = [CategoryResource]
+
+admin.site.register(Project, ProjectAdmin)
+admin.site.register(Category, CategoryAdmin)
